src/pylangbackend/requirements.txt
fastapi
uvicorn[standard]
langchain
python-dotenv
